Moving federated tests into their own module to keep the cache cleaner.
