+++ tests/acceptance/characters/control/test_rpc_control_blockchain.py
@@ -111,60 +111,3 @@ def test_alice_rpc_character_control_grant(alice_rpc_test_client, grant_control_
     assert validate_json_rpc_response_data(response=response,
                                            method_name=method_name,
                                            interface=AliceInterface)
-
-
-def test_bob_rpc_character_control_join_policy(bob_rpc_controller, join_control_request, enacted_federated_policy):
-
-    # Simulate passing in a teacher-uri
-    enacted_federated_policy.bob.remember_node(list(enacted_federated_policy.accepted_ursulas)[0])
-
-    method_name, params = join_control_request
-    request_data = {'method': method_name, 'params': params}
-    response = bob_rpc_controller.send(request_data)
-    assert validate_json_rpc_response_data(response=response,
-                                           method_name=method_name,
-                                           interface=BobInterface)
-
-
-def test_enrico_rpc_character_control_encrypt_message(enrico_rpc_controller_test_client, encrypt_control_request):
-    method_name, params = encrypt_control_request
-    request_data = {'method': method_name, 'params': params}
-    response = enrico_rpc_controller_test_client.send(request_data)
-    assert validate_json_rpc_response_data(response=response,
-                                           method_name=method_name,
-                                           interface=EnricoInterface)
-
-
-def test_bob_rpc_character_control_retrieve(bob_rpc_controller, retrieve_control_request):
-    method_name, params = retrieve_control_request
-    request_data = {'method': method_name, 'params': params}
-    response = bob_rpc_controller.send(request_data)
-    assert validate_json_rpc_response_data(response=response,
-                                           method_name=method_name,
-                                           interface=BobInterface)
-
-
-def test_bob_rpc_character_control_retrieve_with_tmap(
-        enacted_blockchain_policy, blockchain_bob, blockchain_alice,
-        bob_rpc_controller, retrieve_control_request):
-    tmap_64 = b64encode(bytes(enacted_blockchain_policy.treasure_map)).decode()
-    method_name, params = retrieve_control_request
-    params['treasure_map'] = tmap_64
-    request_data = {'method': method_name, 'params': params}
-    response = bob_rpc_controller.send(request_data)
-    assert response.data['result']['cleartexts'][0] == 'Welcome to flippering number 1.'
-
-    # Make a wrong (empty) treasure map
-
-    wrong_tmap = DecentralizedTreasureMap(m=0)
-    wrong_tmap.prepare_for_publication(
-            blockchain_bob.public_keys(DecryptingPower),
-            blockchain_bob.public_keys(SigningPower),
-            blockchain_alice.stamp,
-            b'Wrong!')
-    wrong_tmap._blockchain_signature = b"this is not a signature, but we don't need one for this test....."  # ...because it only matters when Ursula looks at it.
-    tmap_bytes = bytes(wrong_tmap)
-    tmap_64 = b64encode(tmap_bytes).decode()
-    request_data['params']['treasure_map'] = tmap_64
-    with pytest.raises(DecentralizedTreasureMap.IsDisorienting):
-        bob_rpc_controller.send(request_data)
